+++ src/lib/components/bmapv-three-view.vue
@@ -81,7 +81,7 @@ export default {
       this.$children.forEach(component => {
         component.$emit(CONST.MAPV_VIEW_READY_EVENT, this.$bmapComponent);
       });
-      this.bindClick();
+      this.bindEvents();
       this.requestFrame();
     },
     requestFrame() {
@@ -91,9 +91,10 @@ export default {
         this.$bmapComponent.update();
       }
     },
-    bindClick() {
+    bindEvents() {
       this.$bmapComponent.webglLayer.map.map.addEventListener('click', this.clickGltf);
       this.$bmapComponent.webglLayer.map.map.addEventListener('resize', this.resizeCamera);
+      this.$bmapComponent.webglLayer.map.map.addEventListener('mousemove', this.hoverGltf);
       // window.addEventListener('click', this.clickGltf);
     },
     resizeCamera(e) {
@@ -101,7 +102,7 @@ export default {
       camera.aspect = e.size.width / e.size.height;
       camera.updateProjectionMatrix();
     },
-    clickGltf(e) {
+    intersectGltf(e) {
       e = e.domEvent;
       let client = e.srcElement;
       // 通过鼠标点击位置,计算出 raycaster 所需点的位置,以屏幕为中心点,范围 -1 到 1
@@ -122,10 +123,28 @@ export default {
       if (intersects.length > 0) {
         let object = intersects[0];
         // let group = this.getGroup(object.object);
-        let group = object.object.sourceObject;
-        if (group.events && group.events.click) {
-          group.events.click();
-        }
+        return object.object.sourceObject;
+      }
+      return null;
+    },
+    clickGltf(e) {
+      let group = this.intersectGltf(e);
+      if (group.events && group.events.click) {
+        group.events.click();
+      }
+    },
+    hoverGltf(e) {
+      let group = this.intersectGltf(e);
+      if (!group) {
+        this.$bmapComponent.eventObjects.forEach(obj => {
+          if (obj.sourceObject.isHover === true) {
+            obj.sourceObject.isHover = false;
+            obj.sourceObject.events.mouseout();
+          }
+        });
+      } else if (group.events && group.events.mouseover && !group.isHover) {
+        group.isHover = true;
+        group.events.mouseover();
       }
     },
     getGroup(object) {