+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ###v0.0.17
   * 解决GltfThreeLayer图层点击事件触发不稳定问题（该BUG是由于百度的threeLayer图层中修改了camera的projectionMatrix 值，但未对应的修改projectionMatrixInverse 导致的）
+  * GltfThree模型增加mouseover和mouseout事件
 
 ###v0.0.16
   * 修复某些情况下加载模型报错